+++ Estimation/GriddedStreamfunction/@GriddedStreamfunction/decomposeTrajectorySet.m
@@ -33,6 +33,10 @@ mesoscaleTrajectories = trajectories;
 submesoscaleTrajectories = trajectories;
 centeredMesoscaleTrajectories = trajectories;
 centeredSubmesoscaleTrajectories = trajectories;
+cachedTimes = cell(0, 1);
+cachedDegrees = zeros(0, 1);
+cachedBasisMatrices = cell(0, 1);
+cachedKnotPoints = cell(0, 1);
 
 sampleStartIndex = 1;
 for iTrajectory = 1:nTrajectories
@@ -46,14 +50,18 @@ for iTrajectory = 1:nTrajectories
     r0 = sampleData.rAll(sampleIndices(1));
 
     backgroundTrajectories(iTrajectory) = reanchoredBackgroundTrajectory(backgroundTrajectory, ti);
-    mesoscaleTrajectories(iTrajectory) = sampledVelocityTrajectory( ...
-        ti, sampleData.uMesoscaleObserved(sampleIndices), sampleData.vMesoscaleObserved(sampleIndices), componentS, x0, y0);
-    submesoscaleTrajectories(iTrajectory) = sampledVelocityTrajectory( ...
-        ti, sampleData.uSubmesoscaleObserved(sampleIndices), sampleData.vSubmesoscaleObserved(sampleIndices), componentS, 0, 0);
-    centeredMesoscaleTrajectories(iTrajectory) = sampledVelocityTrajectory( ...
-        ti, sampleData.uMesoscaleRelativeObserved(sampleIndices), sampleData.vMesoscaleRelativeObserved(sampleIndices), componentS, q0, r0);
-    centeredSubmesoscaleTrajectories(iTrajectory) = sampledVelocityTrajectory( ...
-        ti, sampleData.uCenteredSubmesoscaleObserved(sampleIndices), sampleData.vCenteredSubmesoscaleObserved(sampleIndices), componentS, 0, 0);
+    [mesoscaleTrajectories(iTrajectory), cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledVelocityTrajectory( ...
+        ti, sampleData.uMesoscaleObserved(sampleIndices), sampleData.vMesoscaleObserved(sampleIndices), componentS, x0, y0, ...
+        cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints);
+    [submesoscaleTrajectories(iTrajectory), cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledVelocityTrajectory( ...
+        ti, sampleData.uSubmesoscaleObserved(sampleIndices), sampleData.vSubmesoscaleObserved(sampleIndices), componentS, 0, 0, ...
+        cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints);
+    [centeredMesoscaleTrajectories(iTrajectory), cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledVelocityTrajectory( ...
+        ti, sampleData.uMesoscaleRelativeObserved(sampleIndices), sampleData.vMesoscaleRelativeObserved(sampleIndices), componentS, q0, r0, ...
+        cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints);
+    [centeredSubmesoscaleTrajectories(iTrajectory), cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledVelocityTrajectory( ...
+        ti, sampleData.uCenteredSubmesoscaleObserved(sampleIndices), sampleData.vCenteredSubmesoscaleObserved(sampleIndices), componentS, 0, 0, ...
+        cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints);
 
     sampleStartIndex = sampleStartIndex + nSamples;
 end
@@ -68,13 +76,12 @@ decomposition = struct( ...
         "submesoscale", centeredSubmesoscaleTrajectories));
 end
 
-function trajectory = sampledVelocityTrajectory(t, uSamples, vSamples, componentS, xMean, yMean)
-xVelocitySpline = InterpolatingSpline.fromGriddedValues(t, uSamples, S=componentS);
-yVelocitySpline = InterpolatingSpline.fromGriddedValues(t, vSamples, S=componentS);
-xIntegralSpline = cumsum(xVelocitySpline);
-yIntegralSpline = cumsum(yVelocitySpline);
-xPathSpline = TensorSpline.fromKnotPoints(xIntegralSpline.knotPoints, xIntegralSpline.xi, S=xIntegralSpline.S, xMean=xMean);
-yPathSpline = TensorSpline.fromKnotPoints(yIntegralSpline.knotPoints, yIntegralSpline.xi, S=yIntegralSpline.S, xMean=yMean);
+function [trajectory, cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledVelocityTrajectory( ...
+        t, uSamples, vSamples, componentS, xMean, yMean, cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints)
+[basisMatrix, knotPoints, cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledSplineBasis( ...
+    t, componentS, cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints);
+velocityCoefficients = basisMatrix \ [reshape(uSamples, [], 1), reshape(vSamples, [], 1)];
+[xPathSpline, yPathSpline] = pathSplinesFromVelocityCoefficients(knotPoints, velocityCoefficients, componentS, xMean, yMean);
 trajectory = TrajectorySpline(t=t, x=xPathSpline, y=yPathSpline);
 end
 
@@ -84,4 +91,30 @@ xSpline = TensorSpline.fromKnotPoints(backgroundTrajectory.x.knotPoints, backgro
 ySpline = TensorSpline.fromKnotPoints(backgroundTrajectory.y.knotPoints, backgroundTrajectory.y.xi, ...
     S=backgroundTrajectory.y.S, xMean=-backgroundTrajectory.y(ti(1)));
 trajectory = TrajectorySpline(t=ti, x=xSpline, y=ySpline);
+end
+
+function [basisMatrix, knotPoints, cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints] = sampledSplineBasis( ...
+        t, componentS, cachedTimes, cachedDegrees, cachedBasisMatrices, cachedKnotPoints)
+t = reshape(t, [], 1);
+for iCache = 1:numel(cachedTimes)
+    if cachedDegrees(iCache) == componentS && isequal(cachedTimes{iCache}, t)
+        basisMatrix = cachedBasisMatrices{iCache};
+        knotPoints = cachedKnotPoints{iCache};
+        return
+    end
+end
+
+knotPoints = BSpline.knotPointsForDataPoints(t, S=componentS);
+basisMatrix = BSpline.matrixForDataPoints(t, knotPoints=knotPoints, S=componentS);
+cachedTimes{end + 1, 1} = t;
+cachedDegrees(end + 1, 1) = componentS;
+cachedBasisMatrices{end + 1, 1} = basisMatrix;
+cachedKnotPoints{end + 1, 1} = knotPoints;
+end
+
+function [xPathSpline, yPathSpline] = pathSplinesFromVelocityCoefficients(knotPoints, velocityCoefficients, splineDegree, xMean, yMean)
+[integratedCoefficients, integratedKnotPoints, integratedS] = BSpline.integratedSplineState( ...
+    velocityCoefficients, knotPoints=knotPoints, S=splineDegree);
+xPathSpline = TensorSpline.fromKnotPoints(integratedKnotPoints, integratedCoefficients(:, 1), S=integratedS, xMean=xMean);
+yPathSpline = TensorSpline.fromKnotPoints(integratedKnotPoints, integratedCoefficients(:, 2), S=integratedS, xMean=yMean);
 end